dummy push

+++ middleware/common/mblue_macro.h
@@ -95,14 +95,14 @@ struct name *type##_##name##_get_instance(void)		\
 
 /* smart pointer macros */
 typedef	void * SMART_DATA;
-#define PB_PAYLOAD(x)			(x && (((uint32_t) (x) & 3) == 0))
-#define INTEGER_P(x)			(((uint32_t) (x) & 3) == 1)
-#define POINTER_S(x)			(GET_POINTER(x) && (((uint32_t) (x) & 3) == 2))	//pointer to static memory
-#define POINTER_H(x)			(GET_POINTER(x) && (((uint32_t) (x) & 3) == 3))	//pointer to heap
-#define GET_INTEGER(x)			((uint32_t) (x) >> 2)
-#define GET_POINTER(x)			((uint32_t) (x) & 0xFFFFFFFC)
-#define MAKE_INTEGER(x)			((void *) (((uint32_t)(x) << 2) | 1))
-#define	MAKE_POINTER_S(x)		((void *) (((uint32_t)(x)) | 2))
+#define PB_PAYLOAD(x)			(x && (((size_t) (x) & 3) == 0))
+#define INTEGER_P(x)			(((size_t) (x) & 3) == 1)
+#define POINTER_S(x)			(GET_POINTER(x) && (((size_t) (x) & 3) == 2))	//pointer to static memory
+#define POINTER_H(x)			(GET_POINTER(x) && (((size_t) (x) & 3) == 3))	//pointer to heap
+#define GET_INTEGER(x)			((size_t) (x) >> 2)
+#define GET_POINTER(x)			((size_t) (x) & 0xFFFFFFFC)
+#define MAKE_INTEGER(x)			((void *) (((size_t)(x) << 2) | 1))
+#define	MAKE_POINTER_S(x)		((void *) (((size_t)(x)) | 2))
 /*#define	MAKE_POINTER_H(x)		((void *) (((uint32_t)(x)) | 3))*/
 
 #endif   /* ----- #ifndef macro_INC  ----- */
